pyproject.toml
[project]
name = "cvassist"
version = "0.1.0"
description = "A CV assistance tool following Clean Architecture principles"
authors = [
    {name = "nauka117", email = "v_kozyrev@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
cvassist = "cvassist.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[tool.poetry]
packages = [{include = "cvassist", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
